Rename files with underscores to files with dashes
Fixes #656
