+++ lib/crash-reporter.js
