+++ lib/discover.js
@@ -8,10 +8,10 @@ var Emitter = events.EventEmitter;
 var debug = require('debug')('discovery');
 
 // Internal
-var lan = require('./lan_connection');
+var lan = require('./lan-connection');
 var logs = require('./logs');
 var Tessel = require('./tessel/tessel');
-var usb = require('./usb_connection');
+var usb = require('./usb-connection');
 
 
 function TesselSeeker() {
