+++ lib/lan-connection.js
