+++ lib/usb-connection.js
@@ -26,7 +26,7 @@ try {
   logs.err('WARNING: No usb controller found on this system.');
 }
 
-var Daemon = require('./usb/usb_daemon');
+var Daemon = require('./usb/usb-daemon');
 
 var TESSEL_VID = 0x1209;
 var TESSEL_PID = 0x7551;
