+++ lib/usb/usb-daemon.js
@@ -6,7 +6,7 @@ var async = require('async');
 var protocol = require('usb-daemon-parser');
 
 // Internal
-var USBProcess = require('./usb_process');
+var USBProcess = require('./usb-process');
 
 var MAX_PROCESS_ID = 255;
