+++ lib/usb/usb-process.js
@@ -5,7 +5,7 @@ var util = require('util');
 
 // Third Party
 var protocol = require('usb-daemon-parser');
-var debug = require('debug')('usb_process');
+var debug = require('debug')('usb-process');
 
 // Internal
 // ...
