+++ test/unit/lan-connection.js
