+++ test/unit/usb-connection.js
